requirements.txt
python-telegram-bot==20.3
openai==0.27.2
schedule==1.1.0
requests==2.26.0
python-dotenv==1.0.0